cleanup stress tests
