+++ tests/stress/stress_test.go
@@ -17,23 +17,13 @@
 package stress
 
 import (
-	"bytes"
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"io/ioutil"
-	"math/rand"
 	"os"
 	"strconv"
 	"testing"
 	"time"
 
-	"github.com/ChainSafe/gossamer/dot/rpc/modules"
-	"github.com/ChainSafe/gossamer/dot/types"
-	"github.com/ChainSafe/gossamer/lib/common"
-	"github.com/ChainSafe/gossamer/lib/common/optional"
-	"github.com/ChainSafe/gossamer/lib/runtime/extrinsic"
-	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/tests/utils"
 
 	log "github.com/ChainSafe/log15"
@@ -41,146 +31,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-var (
-	numNodes   = 3
-	maxRetries = 24
-)
-
 var logger = log.New("pkg", "tests/stress")
-
-// compareChainHeads calls getChainHead for each node in the array
-// it returns a map of chainHead hashes to node key names, and an error if the hashes don't all match
-func compareChainHeads(t *testing.T, nodes []*utils.Node) (map[common.Hash][]string, error) {
-	hashes := make(map[common.Hash][]string)
-	for _, node := range nodes {
-		header := utils.GetChainHead(t, node)
-		logger.Info("got header from node", "hash", header.Hash(), "node", node.Key)
-		hashes[header.Hash()] = append(hashes[header.Hash()], node.Key)
-	}
-
-	var err error
-	if len(hashes) != 1 {
-		err = errors.New("node chain head hashes don't match")
-	}
-
-	return hashes, err
-}
-
-// compareBlocksByNumber calls getBlockByNumber for each node in the array
-// it returns a map of block hashes to node key names, and an error if the hashes don't all match
-func compareBlocksByNumber(t *testing.T, nodes []*utils.Node, num string) (map[common.Hash][]string, error) {
-	hashes := make(map[common.Hash][]string)
-	for _, node := range nodes {
-		hash, err := utils.GetBlockHash(t, node, num)
-		if err != nil {
-			return hashes, err
-		}
-		logger.Info("getting hash from node", "hash", hash, "node", node.Key)
-		hashes[hash] = append(hashes[hash], node.Key)
-	}
-
-	var err error
-	if len(hashes) != 1 {
-		err = errors.New("node hashes don't match")
-	}
-
-	return hashes, err
-}
-
-// compareBlocksByNumberWithRetry calls compareChainHeads, retrying up to maxRetries times if it errors.
-func compareBlocksByNumberWithRetry(t *testing.T, nodes []*utils.Node, num string) {
-	var hashes map[common.Hash][]string
-	var err error
-
-	for i := 0; i < maxRetries; i++ {
-		hashes, err = compareBlocksByNumber(t, nodes, num)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(time.Second)
-	}
-	require.NoError(t, err, hashes)
-}
-
-// compareFinalizedHeads calls getFinalizedHeadByRound for each node in the array
-// it returns a map of finalizedHead hashes to node key names, and an error if the hashes don't all match
-func compareFinalizedHeads(t *testing.T, nodes []*utils.Node) (map[common.Hash][]string, error) {
-	hashes := make(map[common.Hash][]string)
-	for _, node := range nodes {
-		hash := utils.GetFinalizedHead(t, node)
-		logger.Info("got finalized head from node", "hash", hash, "node", node.Key)
-		hashes[hash] = append(hashes[hash], node.Key)
-	}
-
-	var err error
-	if len(hashes) != 1 {
-		err = errors.New("node finalized head hashes don't match")
-	}
-
-	return hashes, err
-}
-
-// compareFinalizedHeadsByRound calls getFinalizedHeadByRound for each node in the array
-// it returns a map of finalizedHead hashes to node key names, and an error if the hashes don't all match
-func compareFinalizedHeadsByRound(t *testing.T, nodes []*utils.Node, round uint64) (map[common.Hash][]string, error) {
-	hashes := make(map[common.Hash][]string)
-	for _, node := range nodes {
-		hash, err := utils.GetFinalizedHeadByRound(t, node, round)
-		if err != nil {
-			continue
-		}
-
-		logger.Info("got finalized head from node", "hash", hash, "node", node.Key, "round", round)
-		hashes[hash] = append(hashes[hash], node.Key)
-	}
-
-	var err error
-	if len(hashes) != 1 {
-		err = errors.New("node finalized head hashes don't match")
-	}
-
-	return hashes, err
-}
-
-// compareChainHeadsWithRetry calls compareChainHeads, retrying up to maxRetries times if it errors.
-func compareChainHeadsWithRetry(t *testing.T, nodes []*utils.Node) {
-	var hashes map[common.Hash][]string
-	var err error
-
-	for i := 0; i < maxRetries; i++ {
-		hashes, err = compareChainHeads(t, nodes)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(time.Second)
-	}
-	require.NoError(t, err, hashes)
-}
-
-// compareFinalizedHeadsWithRetry calls compareFinalizedHeadsByRound, retrying up to maxRetries times if it errors.
-// it returns the finalized hash if it succeeds
-func compareFinalizedHeadsWithRetry(t *testing.T, nodes []*utils.Node, round uint64) common.Hash {
-	var hashes map[common.Hash][]string
-	var err error
-
-	for i := 0; i < maxRetries; i++ {
-		hashes, err = compareFinalizedHeadsByRound(t, nodes, round)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(3 * time.Second)
-	}
-	require.NoError(t, err, fmt.Sprintf("round=%d hashes=%v", round, hashes))
-
-	for h := range hashes {
-		return h
-	}
-
-	return common.Hash{}
-}
 
 func TestMain(m *testing.M) {
 	if utils.MODE != "stress" {
@@ -212,18 +63,24 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func getPendingExtrinsics(t *testing.T, node *utils.Node) [][]byte {
-	respBody, err := utils.PostRPC(utils.AuthorSubmitExtrinsic, utils.NewEndpoint(node.RPCPort), "[]")
+func TestRestartNode(t *testing.T) {
+	nodes, err := utils.InitNodes(numNodes, utils.ConfigDefault)
 	require.NoError(t, err)
 
-	exts := new(modules.PendingExtrinsicsResponse)
-	err = utils.DecodeRPC(t, respBody, exts)
+	err = utils.StartNodes(t, nodes)
 	require.NoError(t, err)
 
-	return *exts
+	errList := utils.TearDown(t, nodes)
+	require.Len(t, errList, 0)
+
+	err = utils.StartNodes(t, nodes)
+	require.NoError(t, err)
+
+	errList = utils.TearDown(t, nodes)
+	require.Len(t, errList, 0)
 }
 
-func TestStressSync(t *testing.T) {
+func TestSync_Basic(t *testing.T) {
 	nodes, err := utils.InitializeAndStartNodes(t, numNodes, utils.GenesisDefault, utils.ConfigDefault)
 	require.NoError(t, err)
 
@@ -244,231 +101,6 @@ func TestStressSync(t *testing.T) {
 
 	errList := utils.TearDown(t, nodes)
 	require.Len(t, errList, 0)
-}
-
-func TestRestartNode(t *testing.T) {
-	nodes, err := utils.InitNodes(numNodes, utils.ConfigDefault)
-	require.NoError(t, err)
-
-	err = utils.StartNodes(t, nodes)
-	require.NoError(t, err)
-
-	errList := utils.TearDown(t, nodes)
-	require.Len(t, errList, 0)
-
-	err = utils.StartNodes(t, nodes)
-	require.NoError(t, err)
-
-	errList = utils.TearDown(t, nodes)
-	require.Len(t, errList, 0)
-
-}
-
-// submitExtrinsicAssertInclusion submits an extrinsic to a random node and asserts that the extrinsic was included in some block
-// and that the nodes remain synced
-func submitExtrinsicAssertInclusion(t *testing.T, nodes []*utils.Node, ext extrinsic.Extrinsic) {
-	tx, err := ext.Encode()
-	require.NoError(t, err)
-
-	txStr := hex.EncodeToString(tx)
-	logger.Info("submitting transaction", "tx", txStr)
-
-	// send extrinsic to random node
-	idx := rand.Intn(len(nodes))
-	prevHeader := utils.GetChainHead(t, nodes[idx]) // get starting header so that we can lookup blocks by number later
-	respBody, err := utils.PostRPC(utils.AuthorSubmitExtrinsic, utils.NewEndpoint(nodes[idx].RPCPort), "\"0x"+txStr+"\"")
-	require.NoError(t, err)
-
-	var hash modules.ExtrinsicHashResponse
-	err = utils.DecodeRPC(t, respBody, &hash)
-	require.Nil(t, err)
-	log.Info("submitted transaction", "hash", hash, "node", nodes[idx].Key)
-	t.Logf("submitted transaction to node %s", nodes[idx].Key)
-
-	// wait for nodes to build block + sync, then get headers
-	time.Sleep(time.Second * 10)
-
-	for i := 0; i < maxRetries; i++ {
-		exts := getPendingExtrinsics(t, nodes[idx])
-		if len(exts) == 0 {
-			break
-		}
-
-		time.Sleep(time.Second)
-	}
-
-	header := utils.GetChainHead(t, nodes[idx])
-	logger.Info("got header from node", "header", header, "hash", header.Hash(), "node", nodes[idx].Key)
-
-	// search from child -> parent blocks for extrinsic
-	var resExts []types.Extrinsic
-	i := 0
-	for header.ExtrinsicsRoot == trie.EmptyHash && i != maxRetries {
-		// check all nodes, since it might have been included on any of the block producers
-		var block *types.Block
-
-		for j := 0; j < len(nodes); j++ {
-			block = utils.GetBlock(t, nodes[j], header.ParentHash)
-			if block == nil {
-				// couldn't get block, increment retry counter
-				i++
-				continue
-			}
-
-			header = block.Header
-			logger.Info("got block from node", "hash", header.Hash(), "node", nodes[j].Key)
-			logger.Debug("got block from node", "header", header, "body", block.Body, "hash", header.Hash(), "node", nodes[j].Key)
-
-			if block.Body != nil && !bytes.Equal(*(block.Body), []byte{0}) {
-				resExts, err = block.Body.AsExtrinsics()
-				require.NoError(t, err, block.Body)
-				break
-			}
-
-			if header.Hash() == prevHeader.Hash() && j == len(nodes)-1 {
-				t.Fatal("could not find extrinsic in any blocks")
-			}
-		}
-
-		if block != nil && block.Body != nil && !bytes.Equal(*(block.Body), []byte{0}) {
-			break
-		}
-	}
-
-	// assert that the extrinsic included is the one we submitted
-	require.Equal(t, 1, len(resExts), "did not find extrinsic in block on any node")
-	require.Equal(t, resExts[0], types.Extrinsic(tx))
-}
-
-func TestStress_IncludeData(t *testing.T) {
-	t.Skip()
-
-	nodes, err := utils.InitializeAndStartNodes(t, numNodes, utils.GenesisDefault, utils.ConfigDefault)
-	require.NoError(t, err)
-
-	defer func() {
-		errList := utils.TearDown(t, nodes)
-		require.Len(t, errList, 0)
-	}()
-
-	time.Sleep(5 * time.Second)
-
-	// create IncludeData extrnsic
-	ext := extrinsic.NewIncludeDataExt([]byte("nootwashere"))
-	submitExtrinsicAssertInclusion(t, nodes, ext)
-
-	compareChainHeadsWithRetry(t, nodes)
-}
-
-func TestStress_StorageChange(t *testing.T) {
-	t.Skip()
-
-	nodes, err := utils.InitializeAndStartNodes(t, numNodes, utils.GenesisDefault, utils.ConfigDefault)
-	require.NoError(t, err)
-
-	defer func() {
-		//TODO: #803 cleanup optimization
-		errList := utils.TearDown(t, nodes)
-		require.Len(t, errList, 0)
-	}()
-
-	time.Sleep(5 * time.Second)
-
-	// create IncludeData extrnsic
-	key := []byte("noot")
-	value := []byte("washere")
-	ext := extrinsic.NewStorageChangeExt(key, optional.NewBytes(true, value))
-	submitExtrinsicAssertInclusion(t, nodes, ext)
-
-	time.Sleep(10 * time.Second)
-
-	// for each node, check that storage was updated accordingly
-	errs := []error{}
-	for _, node := range nodes {
-		logger.Info("getting storage from node", "node", node.Key)
-		res := utils.GetStorage(t, node, key)
-
-		// TODO: why does finalize_block modify the storage value?
-		if bytes.Equal(res, []byte{}) {
-			t.Logf("could not get storage value from node %s", node.Key)
-			errs = append(errs, fmt.Errorf("could not get storage value from node %s\n", node.Key)) //nolint
-		} else {
-			t.Logf("got storage value from node %s: %v", node.Key, res)
-		}
-	}
-
-	require.Equal(t, 0, len(errs), errs)
-	compareChainHeadsWithRetry(t, nodes)
-}
-
-func TestStress_Grandpa_OneAuthority(t *testing.T) {
-	numNodes = 1
-	nodes, err := utils.InitializeAndStartNodes(t, numNodes, utils.GenesisOneAuth, utils.ConfigDefault)
-	require.NoError(t, err)
-
-	defer func() {
-		errList := utils.TearDown(t, nodes)
-		require.Len(t, errList, 0)
-	}()
-
-	time.Sleep(time.Second * 10)
-
-	compareChainHeadsWithRetry(t, nodes)
-	prev, _ := compareFinalizedHeads(t, nodes)
-
-	time.Sleep(time.Second * 10)
-	curr, _ := compareFinalizedHeads(t, nodes)
-	require.NotEqual(t, prev, curr)
-}
-
-func TestStress_Grandpa_ThreeAuthorities(t *testing.T) {
-	numNodes = 3
-	nodes, err := utils.InitializeAndStartNodes(t, numNodes, utils.GenesisThreeAuths, utils.ConfigDefault)
-	require.NoError(t, err)
-
-	defer func() {
-		errList := utils.TearDown(t, nodes)
-		require.Len(t, errList, 0)
-	}()
-
-	time.Sleep(time.Second * 10)
-	fin := compareFinalizedHeadsWithRetry(t, nodes, 1)
-	t.Logf("finalized hash in round 1: %s", fin)
-
-	time.Sleep(time.Second * 10)
-	fin = compareFinalizedHeadsWithRetry(t, nodes, 2)
-	t.Logf("finalized hash in round 2: %s", fin)
-}
-
-func TestStress_Grandpa_NineAuthorities(t *testing.T) {
-	numNodes = 9
-
-	// only log info from 1 node
-	tmpdir, err := ioutil.TempDir("", "gossamer-stress-8")
-	require.NoError(t, err)
-	node, err := utils.RunGossamer(t, numNodes-1, tmpdir, utils.GenesisDefault, utils.ConfigLogGrandpa)
-	require.NoError(t, err)
-
-	// wait and start rest of nodes - if they all start at the same time the first round usually doesn't complete since
-	// all nodes vote for different blocks.
-	time.Sleep(time.Second * 3)
-	nodes, err := utils.InitializeAndStartNodes(t, numNodes-1, utils.GenesisDefault, utils.ConfigLogNone)
-	require.NoError(t, err)
-	nodes = append(nodes, node)
-
-	defer func() {
-		errList := utils.TearDown(t, nodes)
-		require.Len(t, errList, 0)
-	}()
-
-	numRounds := 2
-	for i := 1; i < numRounds+1; i++ {
-		// TODO: this is a long time for a round to complete; this is because syncing is inefficient
-		// need to improve syncing protocol
-		time.Sleep(time.Second * 10)
-		fin := compareFinalizedHeadsWithRetry(t, nodes, uint64(i))
-		t.Logf("finalized hash in round %d: %s", i, fin)
-	}
 }
 
 func TestSync_SingleBlockProducer(t *testing.T) {
